+++ app/dashboard/productos/style.css
@@ -1,0 +1,7 @@
+.productos {
+    font-family: "Ubuntu", sans-serif;
+    font-weight: 700;
+    font-style: normal;
+    color: #C738BD;
+    font-size: 80px;
+}
