+++ app/dashboard/sobre-mi/style.css
@@ -1,0 +1,13 @@
+.sobre-mi {
+    font-family: "Ubuntu", sans-serif;
+    font-weight: 700;
+    font-style: normal;
+    color: #C738BD;
+    font-size: 80px;
+}
+
+.custom-font {
+    font-family: "Ubuntu", sans-serif;
+    font-weight: 300;
+    font-style: normal;
+}
