+++ app/layout.tsx
@@ -1,14 +1,12 @@
-import '@/app/ui/global.css';
-import { inter } from '@/app/ui/fonts';
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+import SideNav from '@/app/ui/dashboard/sidenav';
+ 
+export default function Layout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang="en">
-      <body className={`${inter.className} antialiased`}>{children}</body>
-    </html>
+    <div className="flex h-screen flex-col md:flex-row md:overflow-hidden">
+      <div className="w-full flex-none md:w-64">
+        <SideNav />
+      </div>
+      <div className="flex-grow p-6 md:overflow-y-auto md:p-12">{children}</div>
+    </div>
   );
 }
