+++ app/ui/dashboard/sidenav.tsx
@@ -6,22 +6,17 @@ import { signOut } from '@/auth';
 export default function SideNav() {
   return (
     <div className="flex h-full flex-col px-3 py-4 md:px-2">
-      <Link
-        className="mb-2 flex h-20 items-end justify-start rounded-md bg-purple-400 p-4 md:h-40"
-        href="/"
-      >
-        <div className="w-32 text-white md:w-40">
-          
-        </div>
-      </Link>
+      <a href="https://imgur.com/usjEbcJ" className="mb-2 flex h-20 items-end justify-start rounded-md bg-purple-400 p-4 md:h-40">
+        <img src="https://imgur.com/usjEbcJ" alt="Katty" className="w-40" />
+      </a>
       <div className="flex grow flex-row justify-between space-x-2 md:flex-col md:space-x-0 md:space-y-2">
         <NavLinks />
         <div className="hidden h-auto w-full grow rounded-md bg-gray-50 md:block"></div>
         <form
-         action={async () => {
-          'use server';
-          await signOut();
-        }}
+          action={async () => {
+            'use server';
+            await signOut();
+          }}
         >
           <button className="flex h-[48px] w-full grow items-center justify-center gap-2 rounded-md bg-gray-50 p-3 text-sm font-medium hover:bg-gray-200 hover:text-purple-400 md:flex-none md:justify-start md:p-2 md:px-3">
             <PowerIcon className="w-6" />
