+++ app/ui/fonts.ts
@@ -1,7 +1,7 @@
 import { Inter, Lusitana } from 'next/font/google';
  
-export const inter = Inter({ subsets: ['latin'] });
+/*export const inter = Inter({ subsets: ['latin'] });
 export const lusitana = Lusitana({
     weight: ['400', '700'],
     subsets: ['latin'],
-  });
+  });*/
